[Refactor] 클린 아키텍쳐 구성을 위한 파일 이동
- di모듈의 Module을 제외한 파일 data 모듈로 이동
