+++ data/src/main/kotlin/com/oyj/data/database/BookDatabase.kt
@@ -1,4 +1,4 @@
-package com.oyj.di.database
+package com.oyj.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
