+++ data/src/main/kotlin/com/oyj/data/network/HeaderInterceptor.kt
@@ -1,4 +1,4 @@
-package com.oyj.di.network
+package com.oyj.data.network
 
 import com.oyj.data.BuildConfig
 import okhttp3.Interceptor
